+++ src/main/java/org/jabref/gui/entryeditor/SourceTab.java
@@ -6,16 +6,19 @@ import java.io.StringWriter;
 import java.util.Map;
 import java.util.Objects;
 
+import javafx.beans.property.BooleanProperty;
 import javafx.scene.Node;
 import javafx.scene.control.Tooltip;
 
 import org.jabref.Globals;
+import org.jabref.gui.BasePanel;
 import org.jabref.gui.DialogService;
 import org.jabref.gui.FXDialogService;
 import org.jabref.gui.IconTheme;
 import org.jabref.gui.undo.NamedCompound;
 import org.jabref.gui.undo.UndoableChangeType;
 import org.jabref.gui.undo.UndoableFieldChange;
+import org.jabref.gui.util.DefaultTaskExecutor;
 import org.jabref.logic.bibtex.BibEntryWriter;
 import org.jabref.logic.bibtex.InvalidFieldValueException;
 import org.jabref.logic.bibtex.LatexFieldFormatter;
@@ -23,7 +26,6 @@ import org.jabref.logic.importer.ParserResult;
 import org.jabref.logic.importer.fileformat.BibtexParser;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.model.database.BibDatabase;
-import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.database.BibDatabaseMode;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.InternalBibtexFields;
@@ -41,12 +43,16 @@ public class SourceTab extends EntryEditorTab {
     private static final Log LOGGER = LogFactory.getLog(SourceTab.class);
     private final BibDatabaseMode mode;
     private final BibEntry entry;
+    private final BasePanel panel;
     private CodeArea codeArea;
+    private BooleanProperty movingToDifferentEntry;
 
-    public SourceTab(BibDatabaseContext context, BibEntry entry) {
-        this.mode = context.getMode();
+    public SourceTab(BasePanel panel, BibEntry entry, BooleanProperty movingToDifferentEntry) {
+        this.mode = panel.getBibDatabaseContext().getMode();
         this.entry = entry;
-        context.getDatabase().registerListener(this);
+        this.panel = panel;
+        this.movingToDifferentEntry = movingToDifferentEntry;
+        panel.getBibDatabaseContext().getDatabase().registerListener(this);
         this.setText(Localization.lang("%0 source", mode.getFormattedName()));
         this.setTooltip(new Tooltip(Localization.lang("Show/edit %0 source", mode.getFormattedName())));
         this.setGraphic(IconTheme.JabRefIcon.SOURCE.getGraphicNode());
@@ -81,9 +87,19 @@ public class SourceTab extends EntryEditorTab {
         codeArea.setWrapText(true);
         codeArea.lookup(".styled-text-area").setStyle(
                 "-fx-font-size: " + Globals.prefs.getFontSizeFX() + "pt;");
+        // store source if new tab is selected (if this one is not focused anymore)
         EasyBind.subscribe(codeArea.focusedProperty(), focused -> {
             if (!focused) {
                 storeSource();
+            }
+        });
+
+        // store source if new entry is selected in the maintable and the source tab is focused
+        EasyBind.subscribe(movingToDifferentEntry, newEntrySelected -> {
+            if (newEntrySelected && codeArea.focusedProperty().get()) {
+                DefaultTaskExecutor.runInJavaFXThread(() -> {
+                    storeSource();
+                });
             }
         });
 
@@ -96,6 +112,13 @@ public class SourceTab extends EntryEditorTab {
             codeArea.setEditable(false);
             LOGGER.debug("Incorrect entry", ex);
         }
+
+        // set the database to dirty when something is changed in the source tab
+        EasyBind.subscribe(codeArea.beingUpdatedProperty(), updated -> {
+            if (updated) {
+                panel.markBaseChanged();
+            }
+        });
 
         return new VirtualizedScrollPane<>(codeArea);
     }
@@ -136,8 +159,6 @@ public class SourceTab extends EntryEditorTab {
             NamedCompound compound = new NamedCompound(Localization.lang("source edit"));
             BibEntry newEntry = database.getEntries().get(0);
             String newKey = newEntry.getCiteKeyOptional().orElse(null);
-            boolean entryChanged = false;
-            boolean emptyWarning = (newKey == null) || newKey.isEmpty();
 
             if (newKey != null) {
                 entry.setCiteKey(newKey);
@@ -154,7 +175,6 @@ public class SourceTab extends EntryEditorTab {
                     compound.addEdit(
                             new UndoableFieldChange(entry, fieldName, fieldValue, null));
                     entry.clearField(fieldName);
-                    entryChanged = true;
                 }
             }
 
@@ -170,7 +190,6 @@ public class SourceTab extends EntryEditorTab {
 
                     compound.addEdit(new UndoableFieldChange(entry, fieldName, oldValue, newValue));
                     entry.setField(fieldName, newValue);
-                    entryChanged = true;
                 }
             }
 
@@ -178,23 +197,9 @@ public class SourceTab extends EntryEditorTab {
             if (!Objects.equals(newEntry.getType(), entry.getType())) {
                 compound.addEdit(new UndoableChangeType(entry, entry.getType(), newEntry.getType()));
                 entry.setType(newEntry.getType());
-                entryChanged = true;
             }
             compound.end();
 
-            // TODO: Add undo
-            //panel.getUndoManager().addEdit(compound);
-
-            // TODO: Warn about duplicate/empty bibtext key
-            /*
-            if (panel.getDatabase().getDuplicationChecker().isDuplicateCiteKeyExisting(entry)) {
-                warnDuplicateBibtexkey();
-            } else if (emptyWarning) {
-                warnEmptyBibtexkey();
-            } else {
-                panel.output(Localization.lang("Stored entry") + '.');
-            }
-            */
         } catch (InvalidFieldValueException | IOException ex) {
             // The source couldn't be parsed, so the user is given an
             // error message, and the choice to keep or revert the contents